+++ acvm/src/compiler/mod.rs
@@ -15,8 +15,6 @@ mod transformers;
 use optimizers::{GeneralOptimizer, RangeOptimizer};
 use transformers::{CSatTransformer, FallbackTransformer, R1CSTransformer};
 
-pub use optimizers::{CircuitSimplifier, SimplifyResult};
-
 #[derive(PartialEq, Eq, Debug, Error)]
 pub enum CompileError {
     #[error("The blackbox function {0} is not supported by the backend and acvm does not have a fallback implementation")]
@@ -28,7 +26,6 @@ pub fn compile(
     acir: Circuit,
     np_language: Language,
     is_opcode_supported: impl Fn(&Opcode) -> bool,
-    simplifier: &CircuitSimplifier,
 ) -> Result<(Circuit, Vec<OpcodeLabel>), CompileError> {
     // Instantiate the optimizer.
     // Currently the optimizer and reducer are one in the same
@@ -40,7 +37,7 @@ pub fn compile(
 
     // Fallback transformer pass
     let (acir, opcode_label) =
-        FallbackTransformer::transform(acir, is_opcode_supported, simplifier, opcode_labels)?;
+        FallbackTransformer::transform(acir, is_opcode_supported, opcode_labels)?;
 
     // General optimizer pass
     let mut opcodes: Vec<Opcode> = Vec::new();
@@ -65,8 +62,8 @@ pub fn compile(
         }
         crate::Language::PLONKCSat { width } => {
             let mut csat = CSatTransformer::new(*width);
-            for value in &acir.inputs {
-                csat.solvable(*value);
+            for value in acir.circuit_arguments() {
+                csat.mark_solvable(value);
             }
             csat
         }
@@ -117,7 +114,7 @@ pub fn compile(
                 match func {
                     acir::circuit::opcodes::BlackBoxFuncCall::AND { output, .. }
                     | acir::circuit::opcodes::BlackBoxFuncCall::XOR { output, .. } => {
-                        transformer.solvable(*output)
+                        transformer.mark_solvable(*output)
                     }
                     acir::circuit::opcodes::BlackBoxFuncCall::RANGE { .. } => (),
                     acir::circuit::opcodes::BlackBoxFuncCall::SHA256 { outputs, .. }
@@ -132,7 +129,7 @@ pub fn compile(
                     }
                     | acir::circuit::opcodes::BlackBoxFuncCall::Blake2s { outputs, .. } => {
                         for witness in outputs {
-                            transformer.solvable(*witness);
+                            transformer.mark_solvable(*witness);
                         }
                     }
                     acir::circuit::opcodes::BlackBoxFuncCall::FixedBaseScalarMul {
@@ -140,8 +137,8 @@ pub fn compile(
                         ..
                     }
                     | acir::circuit::opcodes::BlackBoxFuncCall::Pedersen { outputs, .. } => {
-                        transformer.solvable(outputs.0);
-                        transformer.solvable(outputs.1)
+                        transformer.mark_solvable(outputs.0);
+                        transformer.mark_solvable(outputs.1)
                     }
                     acir::circuit::opcodes::BlackBoxFuncCall::HashToField128Security {
                         output,
@@ -150,7 +147,7 @@ pub fn compile(
                     | acir::circuit::opcodes::BlackBoxFuncCall::EcdsaSecp256k1 { output, .. }
                     | acir::circuit::opcodes::BlackBoxFuncCall::EcdsaSecp256r1 { output, .. }
                     | acir::circuit::opcodes::BlackBoxFuncCall::SchnorrVerify { output, .. } => {
-                        transformer.solvable(*output)
+                        transformer.mark_solvable(*output)
                     }
                 }
 
@@ -158,23 +155,22 @@ pub fn compile(
                 transformed_gates.push(opcode.clone());
             }
             Opcode::Directive(directive) => {
-                dbg!(&directive);
                 match directive {
                     Directive::Invert { result, .. } => {
-                        transformer.solvable(*result);
+                        transformer.mark_solvable(*result);
                     }
                     Directive::Quotient(quotient_directive) => {
-                        transformer.solvable(quotient_directive.q);
-                        transformer.solvable(quotient_directive.r);
+                        transformer.mark_solvable(quotient_directive.q);
+                        transformer.mark_solvable(quotient_directive.r);
                     }
                     Directive::ToLeRadix { b, .. } => {
                         for w in b {
-                            transformer.solvable(*w);
+                            transformer.mark_solvable(*w);
                         }
                     }
                     Directive::PermutationSort { bits, .. } => {
                         for w in bits {
-                            transformer.solvable(*w);
+                            transformer.mark_solvable(*w);
                         }
                     }
                     Directive::Log(_) => (),
@@ -188,10 +184,10 @@ pub fn compile(
             Opcode::Brillig(brillig) => {
                 for output in &brillig.outputs {
                     match output {
-                        BrilligOutputs::Simple(w) => transformer.solvable(*w),
+                        BrilligOutputs::Simple(w) => transformer.mark_solvable(*w),
                         BrilligOutputs::Array(v) => {
                             for w in v {
-                                transformer.solvable(*w);
+                                transformer.mark_solvable(*w);
                             }
                         }
                     }
@@ -208,9 +204,9 @@ pub fn compile(
             current_witness_index,
             opcodes: transformed_gates,
             // The optimizer does not add new public inputs
+            private_parameters: acir.private_parameters,
             public_parameters: acir.public_parameters,
             return_values: acir.return_values,
-            inputs: acir.inputs,
         },
         new_opcode_labels,
     ))