+++ acvm/src/compiler/transformers/fallback.rs
@@ -1,5 +1,3 @@
-use crate::compiler::optimizers::CircuitSimplifier;
-
 use super::super::CompileError;
 use acir::{
     circuit::{opcodes::BlackBoxFuncCall, Circuit, Opcode, OpcodeLabel},
@@ -15,49 +13,42 @@ impl FallbackTransformer {
     pub(crate) fn transform(
         acir: Circuit,
         is_supported: impl Fn(&Opcode) -> bool,
-        simplifier: &CircuitSimplifier,
         opcode_labels: Vec<OpcodeLabel>,
     ) -> Result<(Circuit, Vec<OpcodeLabel>), CompileError> {
         let mut acir_supported_opcodes = Vec::with_capacity(acir.opcodes.len());
         let mut new_opcode_labels = Vec::with_capacity(opcode_labels.len());
         let mut witness_idx = acir.current_witness_index + 1;
-        // add opcodes for defining the witness that will be solved through simplification but must be kept
-        for w in &simplifier.defined {
-            acir_supported_opcodes.push(simplifier.define(w));
-        }
+
         for (idx, opcode) in acir.opcodes.into_iter().enumerate() {
-            if !simplifier.solved_gates.contains(&idx) {
-                match &opcode {
-                    Opcode::Arithmetic(_)
-                    | Opcode::Directive(_)
-                    | Opcode::Brillig(_)
-                    | Opcode::Block(_)
-                    | Opcode::ROM(_)
-                    | Opcode::RAM(_) => {
-                        // directive, arithmetic expression or blocks are handled by acvm
+            match &opcode {
+                Opcode::Arithmetic(_)
+                | Opcode::Directive(_)
+                | Opcode::Brillig(_)
+                | Opcode::Block(_)
+                | Opcode::ROM(_)
+                | Opcode::RAM(_) => {
+                    // directive, arithmetic expression or blocks are handled by acvm
+                    new_opcode_labels.push(opcode_labels[idx]);
+                    acir_supported_opcodes.push(opcode);
+                    continue;
+                }
+                Opcode::BlackBoxFuncCall(bb_func_call) => {
+                    // We know it is an black box function. Now check if it is
+                    // supported by the backend. If it is supported, then we can simply
+                    // collect the opcode
+                    if is_supported(&opcode) {
                         new_opcode_labels.push(opcode_labels[idx]);
                         acir_supported_opcodes.push(opcode);
                         continue;
-                    }
-                    Opcode::BlackBoxFuncCall(bb_func_call) => {
-                        // We know it is an black box function. Now check if it is
-                        // supported by the backend. If it is supported, then we can simply
-                        // collect the opcode
-                        if is_supported(&opcode) {
-                            new_opcode_labels.push(opcode_labels[idx]);
-                            acir_supported_opcodes.push(opcode);
-                            continue;
-                        } else {
-                            // If we get here then we know that this black box function is not supported
-                            // so we need to replace it with a version of the opcode which only uses arithmetic
-                            // expressions
-                            let (updated_witness_index, opcodes_fallback) =
-                                Self::opcode_fallback(bb_func_call, witness_idx)?;
-                            witness_idx = updated_witness_index;
-                            new_opcode_labels
-                                .extend(vec![opcode_labels[idx]; opcodes_fallback.len()]);
-                            acir_supported_opcodes.extend(opcodes_fallback);
-                        }
+                    } else {
+                        // If we get here then we know that this black box function is not supported
+                        // so we need to replace it with a version of the opcode which only uses arithmetic
+                        // expressions
+                        let (updated_witness_index, opcodes_fallback) =
+                            Self::opcode_fallback(bb_func_call, witness_idx)?;
+                        witness_idx = updated_witness_index;
+                        new_opcode_labels.extend(vec![opcode_labels[idx]; opcodes_fallback.len()]);
+                        acir_supported_opcodes.extend(opcodes_fallback);
                     }
                 }
             }
@@ -67,9 +58,9 @@ impl FallbackTransformer {
             Circuit {
                 current_witness_index: witness_idx,
                 opcodes: acir_supported_opcodes,
+                private_parameters: acir.private_parameters,
                 public_parameters: acir.public_parameters,
                 return_values: acir.return_values,
-                inputs: acir.inputs,
             },
             new_opcode_labels,
         ))
@@ -85,7 +76,7 @@ impl FallbackTransformer {
                     lhs.num_bits, rhs.num_bits,
                     "number of bits specified for each input must be the same"
                 );
-                stdlib::fallback::and(
+                stdlib::blackbox_fallbacks::and(
                     Expression::from(lhs.witness),
                     Expression::from(rhs.witness),
                     *output,
@@ -98,7 +89,7 @@ impl FallbackTransformer {
                     lhs.num_bits, rhs.num_bits,
                     "number of bits specified for each input must be the same"
                 );
-                stdlib::fallback::xor(
+                stdlib::blackbox_fallbacks::xor(
                     Expression::from(lhs.witness),
                     Expression::from(rhs.witness),
                     *output,
@@ -108,9 +99,37 @@ impl FallbackTransformer {
             }
             BlackBoxFuncCall::RANGE { input } => {
                 // Note there are no outputs because range produces no outputs
-                stdlib::fallback::range(
+                stdlib::blackbox_fallbacks::range(
                     Expression::from(input.witness),
                     input.num_bits,
+                    current_witness_idx,
+                )
+            }
+            #[cfg(feature = "unstable-fallbacks")]
+            BlackBoxFuncCall::SHA256 { inputs, outputs } => {
+                let mut sha256_inputs = Vec::new();
+                for input in inputs.iter() {
+                    let witness_index = Expression::from(input.witness);
+                    let num_bits = input.num_bits;
+                    sha256_inputs.push((witness_index, num_bits));
+                }
+                stdlib::blackbox_fallbacks::sha256(
+                    sha256_inputs,
+                    outputs.to_vec(),
+                    current_witness_idx,
+                )
+            }
+            #[cfg(feature = "unstable-fallbacks")]
+            BlackBoxFuncCall::Blake2s { inputs, outputs } => {
+                let mut blake2s_input = Vec::new();
+                for input in inputs.iter() {
+                    let witness_index = Expression::from(input.witness);
+                    let num_bits = input.num_bits;
+                    blake2s_input.push((witness_index, num_bits));
+                }
+                stdlib::blackbox_fallbacks::blake2s(
+                    blake2s_input,
+                    outputs.to_vec(),
                     current_witness_idx,
                 )
             }